AI coding Assistant web app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
python-dotenv==1.0.1
langchain-google-genai>=2.1.0
langgraph>=0.4.0
langchain-core>=0.3.0
duckduckgo-search>=7.0.0
